feat(article): add more highlight to code (plain, c, c++, julia, py, r, rust), refacto(article): file architecture of article code
